moved config setup and management back into the sparkpost object
