+++ lib/SparkPost/APIResource.php
@@ -1,8 +1,7 @@
 <?php
 namespace SparkPost;
-use Ivory\HttpAdapter\Configuration;
-use Ivory\HttpAdapter\HttpAdapterInterface;
 use Ivory\HttpAdapter\HttpAdapterException;
+use SparkPost\SparkPost;
 
 /**
  * @desc SDK interface for managing SparkPost API endpoints
@@ -29,110 +28,13 @@ class APIResource {
 	 */
 	protected static $structure = [];
 
-  /**
-   * @dec connection config for making requests.
-   */
-  private $config;
-
-  /**
-   * @desc Ivory\HttpAdapter\HttpAdapterInterface to make requests through.
-   */
-  protected $httpAdapter;
-
-  /**
-   * @desc Default config values. Passed in values will override these.
-   */
-  private static $apiDefaults = [
-		'host'=>'api.sparkpost.com',
-		'protocol'=>'https',
-		'port'=>443,
-		'strictSSL'=>true,
-		'key'=>'',
-		'version'=>'v1'
-	];
-
 	/**
 	 * @desc Initializes config and httpAdapter for use later.
-   * @param $httpAdapter Ivory\HttpAdapter\HttpAdapterInterface to make requests through.
-   * @param $config connection config for making requests.
+   * @param $sparkpost SparkPost\SparkPost provides api configuration information
 	 */
-	public function __construct($httpAdapter, $config) {
-    //config needs to be setup before adapter because of default adapter settings
-    $this->setConfig($config);
-    $this->setHttpAdapter($httpAdapter);
+	public function __construct(SparkPost $sparkpost) {
+    $this->sparkpost = $sparkpost;
   }
-
-  /**
-	 * @desc Merges passed in headers with default headers for http requests
-	 * @return Array - headers to be set on http requests
-	 */
-	private function getHttpHeaders(Array $headers = null) {
-		$defaultOptions = [
-			'Authorization' => $this->config['key'],
-			'Content-Type' => 'application/json',
-		];
-
-		// Merge passed in headers with defaults
-		if (!is_null($headers)) {
-			foreach ($headers as $header => $value) {
-				$defaultOptions[$header] = $value;
-			}
-		}
-		return $defaultOptions;
-	}
-
-
-  /**
-	 * @desc Helper function for getting the configuration for http requests
-	 * @return \Ivory\HttpAdapter\Configuration
-	 */
-  private function getHttpConfig($config) {
-		// get composer.json to extract version number
-		$composerFile = file_get_contents(dirname(__FILE__) . "/../../composer.json");
-		$composer = json_decode($composerFile, true);
-
-		// create Configuration for http adapter
-		$httpConfig = new Configuration();
-		$baseUrl = $config['protocol'] . '://' . $config['host'] . ($config['port'] ? ':' . $config['port'] : '') . '/api/' . $config['version'];
-		$httpConfig->setBaseUri($baseUrl);
-		$httpConfig->setUserAgent('php-sparkpost/' . $composer['version']);
-		return $httpConfig;
-	}
-
-  /**
-   * @desc Validates and sets up the httpAdapter
-   * @param $httpAdapter Ivory\HttpAdapter\HttpAdapterInterface to make requests through.
-   * @throws \Exception
-   */
-  public function setHttpAdapter($httpAdapter) {
-  	if (!$httpAdapter instanceOf HttpAdapterInterface) {
-			throw new \Exception('$httpAdapter paramter must be a valid Ivory\HttpAdapter');
-		}
-
-    $this->httpAdapter = $httpAdapter;
-    $this->httpAdapter->setConfiguration($this->getHttpConfig($this->config));
-  }
-
-	/**
-	 * Allows the user to pass in values to override the defaults and set their API key
-	 * @param Array $settingsConfig - Hashmap that contains config values for the SDK to connect to SparkPost
-	 * @throws \Exception
-	 */
-	public function setConfig(Array $settingsConfig) {
-		// Validate API key because its required
-    if (!isset($settingsConfig['key']) || empty(trim($settingsConfig['key']))){
-			throw new \Exception('You must provide an API key');
-    }
-
-		$this->config = self::$apiDefaults;
-
-    // set config, overriding defaults
-		foreach ($settingsConfig as $configOption => $configValue) {
-			if(key_exists($configOption, $this->config)) {
-				$this->config[$configOption] = $configValue;
-			}
-		}
-	}
 
 	/**
 	 * @desc Private Method helper to reference parameter mappings and set the right value for the right parameter
@@ -282,7 +184,7 @@ class APIResource {
 
 		//make request
 		try {
-			$response = $this->httpAdapter->send($url, $action, $this->getHttpHeaders(), $body);
+			$response = $this->sparkpost->httpAdapter->send($url, $action, $this->sparkpost->getHttpHeaders(), $body);
 			return json_decode($response->getBody()->getContents(), true);
 		}
 		/*
